requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.106
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
openai==0.27.1
packaging==23.0
protobuf==3.20.3
pycodestyle==2.10.0
pydantic==1.10.6
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.46
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.14
yarl==1.8.2
